style: рефакторинг app
