+++ middlewares/errorsHandler.js
@@ -1,0 +1,10 @@
+const errorsHandler = (err, req, res, next) => {
+  const { statusCode = 500, message } = err;
+
+  res.status(statusCode).send({
+    message: statusCode === 500 ? 'На сервере произошла ошибка' : message,
+  });
+  next();
+};
+
+module.exports = { errorsHandler };
